+++ include/ros_phoenix/base_component.hpp
@@ -30,7 +30,7 @@ namespace ros_phoenix
         {
             this->declare_parameter<std::string>("interface", "can0");
             ctre::phoenix::platform::can::SetCANInterface(this->get_parameter("interface").as_string().c_str());
-            //c_SetPhoenixDiagnosticsStartTime(-1); // disable diag server
+            c_SetPhoenixDiagnosticsStartTime(-1); // disable diag server
 
             this->declare_parameter<int>("id", 0);
             this->declare_parameter<int>("period_ms", 20);
