+++ launch/demo-nodes.launch.py
@@ -1,0 +1,34 @@
+from launch import LaunchDescription
+from launch_ros.actions import Node
+
+
+def generate_launch_description():
+    """Generate launch description with multiple nodes."""
+    return LaunchDescription(
+        [
+            Node(
+                package="ros_phoenix",
+                executable="talon_node",
+                name="front_left",
+                parameters=[{"id": 15, "invert": True}],
+            ),
+            Node(
+                package="ros_phoenix",
+                executable="talon_node",
+                name="front_right",
+                parameters=[{"id": 18, "invert": False}],
+            ),
+            Node(
+                package="ros_phoenix",
+                executable="talon_node",
+                name="back_left",
+                parameters=[{"id": 10, "invert": True}],
+            ),
+            Node(
+                package="ros_phoenix",
+                executable="talon_node",
+                name="back_right",
+                parameters=[{"id": 19, "invert": False}],
+            ),
+        ]
+    )
